--- capitalizeDecorator.rb
@@ -1,9 +1,0 @@
-require_relative './baseDecorator'
-
-
-class CapitalizeDecorator < Decorator
-    def correct_name
-        @nameable_object.correct_name.upcase
-    end
-end
-
