+++ capitalize_decorator.rb
@@ -1,0 +1,7 @@
+require_relative './base_decorator'
+
+class CapitalizeDecorator < Decorator
+  def correct_name
+    @nameable_object.correct_name.upcase
+  end
+end
