+++ person.rb
@@ -1,12 +1,12 @@
 require_relative './nameable'
 
-
 class Person < Nameable
   def initialize(age, name = 'Unknown', parent_permission: true)
     @id = Random.rand(1..1000)
     @name = name
     @parent_permission = parent_permission
     @age = age
+    super()
   end
 
   def correct_name
@@ -26,5 +26,3 @@ class Person < Nameable
   attr_reader :id
   attr_accessor :name, :age
 end
-
-
