--- trimmerDecorator.rb
@@ -1,7 +1,0 @@
-require_relative './baseDecorator'
-
-class TrimmerDecorator < Decorator
-    def correct_name
-        @nameable_object.correct_name[0..9]
-    end
-end
