+++ trimmer_decorator.rb
@@ -1,0 +1,7 @@
+require_relative './base_decorator'
+
+class TrimmerDecorator < Decorator
+  def correct_name
+    @nameable_object.correct_name[0..9]
+  end
+end
